+++ docarray/typing/tensor/__init__.py
@@ -2,6 +2,13 @@ from docarray.typing.tensor.embedding import Embedding, NdArrayEmbedding
 from docarray.typing.tensor.ndarray import NdArray
 from docarray.typing.tensor.tensor import Tensor
 
+__all__ = [
+    'NdArray',
+    'Tensor',
+    'Embedding',
+    'NdArrayEmbedding',
+]
+
 try:
     import torch  # noqa: F401
 except ImportError:
@@ -10,10 +17,4 @@ else:
     from docarray.typing.tensor.embedding import TorchEmbedding  # noqa: F401
     from docarray.typing.tensor.torch_tensor import TorchTensor  # noqa: F401
 
-
-__all__ = [
-    'NdArray',
-    'Tensor',
-    'Embedding',
-    'NdArrayEmbedding',
-]
+    __all__.extend(['TorchEmbedding', 'TorchTensor'])
